+++ samples/src/test/groovy/com/tkruse/gradle/groovysh/SimpleGroovySpecification.groovy
@@ -33,6 +33,6 @@ class SimpleGroovySpecification extends Specification {
         assert !bytesOut.toString().contains('Exception')
 
         where:
-        groovyVersion << ['2.0.0', '2.0.8', '2.1.0', '2.1.9', '2.2.0', '2.2.1', '2.2.2', '2.3.3']
+        groovyVersion << ['2.0.0', '2.0.8', '2.1.0', '2.1.9', '2.2.0', '2.2.1', '2.2.2', '2.3.0', '2.3.4']
     }
 }