+++ src/test/groovy/com/tkruse/gradle/groovysh/ApplicationShellTaskTest.groovy
@@ -41,7 +41,7 @@ class ApplicationShellTaskTest {
 
     @Test
     void testGroovyVersions() {
-        for (String version in ['2.2.1', '2.2.2', '2.3.0', '2.3.3']) {
+        for (String version in ['2.2.1', '2.2.2', '2.3.0', '2.3.4']) {
             Project project = TestHelper.createProjectWithPlugin()
             project.dependencies.add('testCompile', 'junit:junit-dep:4.11')
 
@@ -87,7 +87,7 @@ class ApplicationShellTaskTest {
                 project.configurations.appShellConf_shell.dependencies.asList().collect { it.name + it.version }
         assert dependencyVersions.contains('jline2.11')
         assert dependencyVersions.contains('commons-cli1.2')
-        assert dependencyVersions.contains('groovy-all2.3.3')
+        assert dependencyVersions.contains('groovy-all2.3.4')
         assert shellTask.dependsOn.contains('classes')
         assert shellTask.classpath.asPath.contains('groovy')
         assert shellTask.classpath.asPath.contains('junit')