+++ src/test/groovy/com/tkruse/gradle/groovysh/BuildShellTaskTest.groovy
@@ -31,7 +31,7 @@ class BuildShellTaskTest {
 
     @Test
     void testGroovyVersions() {
-        for (String version in ['2.2.1', '2.2.2', '2.3.0', '2.3.3']) {
+        for (String version in ['2.2.1', '2.2.2', '2.3.0', '2.3.4']) {
             for (String gradleVersion in ['1.9', '1.10', '1.11', '1.12', '2.0']) {
                 Project project = TestHelper.createProjectWithPlugin()
                 project.dependencies.add('testCompile', 'junit:junit-dep:4.11')
@@ -69,7 +69,7 @@ class BuildShellTaskTest {
                 project.configurations.buildShellConf_buildShell.dependencies.asList().collect { it.name + it.version }
         assert dependencyVersions.contains('jline2.11')
         assert dependencyVersions.contains('commons-cli1.2')
-        assert dependencyVersions.contains('groovy-all2.3.3')
+        assert dependencyVersions.contains('groovy-all2.3.4')
         assert dependencyVersions.contains('gradle-tooling-api2.0')
 
     }